requirements.txt
pyrotgfork==2.2.16
tgcrypto-pyrofork==1.2.7
aiogram==3.22.0
asyncio==4.0.0
aiohttp==3.12.15
configparser==7.2.0
aiogram_i18n==1.4
fluent_compiler==1.1
fluent.runtime==0.4.0
python-dotenv==1.1.1